doc: updated broken links (#88)
* Replaced broken link to medium article with the kdnuggets article.

* Fixed links to images and links to files.

I did not update the step 3 and 4 since I'm not certain how you all want to deal with them. There are missing files and links may need to be updated. And the paragraphs may need to be reworded due to structure change. The project structure section also needs to be updated.

* Updated link to cookiecutter readme.